Change 1: Implement modular pipeline set algebra: add new module and tests for namespace boundary calculations and dataset ownership
